Capsule specific image check status return
A change is made to return true always even if the
uploaded capsule fails the HASH and signature checks.
Which was intended for Seamless update capsules, expecting
Seamless Capsule fixes their image signing and format
issues soon.
Now, we made the status conditional as we need the correct
status to be returned for PFR capsules.

Tested: tested with corrupted PFR capsules.

Change-Id: Ib3e771ea2bc506340cfd90496af36655b1bf5a34
Signed-off-by: Vikram Bodireddy <vikram.bodireddy@intel.com>